requirements.txt
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
fastapi-jwt-auth==0.5.0
filelock==3.8.0
flake8==5.0.4
greenlet==1.1.3
h11==0.13.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
jwt==1.3.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pipenv==2022.10.12
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
PyJWT==1.7.1
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
redis==4.3.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
validators==0.20.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
